refine paranoid_confirm:Autoall
If confirmation for menustyle:Full 'A' choice is on and player
chooses 'A' but answers no at the "Really autoselect All?" prompt,
just remove 'A' from the list of choices and act on the rest.  If it
was the only choice, the menu operation will behave as if no choices
had been made.  (That was the previous behavior even when other
choices were also present.)  But when other choices have been made
along with 'A', use them without 'A':  the normal second menu will
be put up, listing objects matching the specified classes and not
autoselecting anything.