mondead() uncluttering
Pull request #1005 introduced some code that needed minor reformatting.
This goes beyond that, splitting the livelog/#chronicle handling and
the vampire unshifting code out of mondead() into separate routines.

The livelog phrasing for death of a unique monster always assumed that
the hero was directly responsible:
 "killed|destroyed <unique mon|shk+details> (Nth time)."
If it dies while monsters are moving, switch to:
 "<unique mon|shk+details> has been killed|destroyed (Nth time)."
It isn't phrased as if the hero is responsible if the monster dies due
to passive counter-attack against poly'd hero but that isn't blatantly
wrong and doesn't seem worth worrying about.