documentation for PR #1071 - highlight criticalhp
Pull request from vultur-cadens:  add a new status highlight pattern
OPTIONS=hilite:hitpoints/criticalhp/color&attribute
to highlight hit points when they drop to the point where prayer
classifies them as a major problem.  The threshold for that varies
depending on max HP and experience level.

It affects hitpointbar as well as the status field for hit points.

Right now a highlight rule for HP down or HP changed or HP up takes
precedence over criticalhp until the temporary highlight times out.
I'm not sure how best to deal with that.  With regeneration and an
up or changed rule for HP, the criticalhp highlight probably won't be
seen even with a very short statushilites timeout.

I've added a fixes entry and updated Guidebook.mn but Guidebook.tex
is lagging.

Closes #1071