data.base entry for stethoscope
I was going to add a joke entry for "microscope" to data.base and
disocovered that stethoscope was missing.  (There is a rumor about
cursed ones.)  Add a lame entry for it so that the joke has something
to relate to.

Searching for "stethoscope" led to the "cope" comment, which is near
the "cornuthaum" entry.  Add a comment giving Janet credit for coining
the term.