github pull request #1108 - new data.base entries
Pull request from NullCGT:  a lot of new data.base entries.  Over 20
separate commits; this time I didn't squash them together because I
didn't want to mess with their log messages.

Three had non-ascii characters where apostrophe or single-quote or
long dash are intended.  _The_Invisible_Man_ for stalker was the worst
offender.  While fixing those, I changed period followed by single
space when separating sentences to period followed by two spaces but
didn't do that for the other new entires.  A few had trailing spaces and
one existing entry had trailing tabs.  I've fixed up those situations.

The ice box quote is out of whack with everything else due to wider
lines but the capitalization indicates that it's from a poem that was
formatted that way.  I left it as is instead of splitting several of
the lines.

The conical hat entry is iffy since nethack's conical hat is not a
wide-brimmed witch's hat, but I've left it in.  Likewise helm of
brilliance says that it's made of steel but nethack's is now crystal.

Closes #1108